+++ src/swap/Marketplace.sol
@@ -9,20 +9,15 @@ contract MarketPlace {
     mapping(uint256 => Order) public ercErcOrders; // cant have struct as key, nor tupple
 
     event placeOrder(
-        uint indexed orderId,
+        uint256 indexed orderId,
         address indexed requesterAddress,
         uint256 offeringAmount,
         address offeringToken,
         uint256 askingAmount,
         address askingToken
     );
-    event acceptOrder(
-        uint indexed orderId,
-        address indexed who,
-        uint buyAmount,
-        uint saleAmount
-    );
-    event withdrawOrder(uint indexed orderId);
+    event acceptOrder(uint256 indexed orderId, address indexed who, uint256 buyAmount, uint256 saleAmount);
+    event withdrawOrder(uint256 indexed orderId);
 
     uint256 public nextOrderId;
 
@@ -34,23 +29,16 @@ contract MarketPlace {
         address requesterAddress;
     }
 
-    function placeErcErcOrder(
-        address sellingToken,
-        uint saleAmount,
-        address buyingToken,
-        uint buyAmount
-    ) public {
+    function placeErcErcOrder(address sellingToken, uint256 saleAmount, address buyingToken, uint256 buyAmount)
+        public
+    {
         require(sellingToken != address(0x0));
         require(buyingToken != address(0x0));
 
         // "lock" selling token by transferring to contract
-        IERC20(sellingToken).safeTransferFrom(
-            msg.sender,
-            address(this),
-            saleAmount
-        );
+        IERC20(sellingToken).safeTransferFrom(msg.sender, address(this), saleAmount);
 
-        uint id = nextOrderId++;
+        uint256 id = nextOrderId++;
         Order memory order = Order({
             offeringAmount: saleAmount,
             offeringToken: sellingToken,
@@ -62,16 +50,11 @@ contract MarketPlace {
         ercErcOrders[id] = order;
 
         emit placeOrder(
-            id,
-            order.requesterAddress,
-            order.offeringAmount,
-            order.offeringToken,
-            order.askingAmount,
-            order.askingToken
+            id, order.requesterAddress, order.offeringAmount, order.offeringToken, order.askingAmount, order.askingToken
         );
     }
 
-    function acceptErcErcOrder(uint id, uint saleAmount) public {
+    function acceptErcErcOrder(uint256 id, uint256 saleAmount) public {
         Order memory order = ercErcOrders[id];
 
         // make sure the storage item exists
@@ -79,8 +62,7 @@ contract MarketPlace {
 
         require(saleAmount <= order.askingAmount);
         // todo: make this safe
-        uint buyAmount = (saleAmount * order.offeringAmount) /
-            order.askingAmount;
+        uint256 buyAmount = (saleAmount * order.offeringAmount) / order.askingAmount;
 
         assert(saleAmount > 0);
         assert(buyAmount > 0);
@@ -89,24 +71,17 @@ contract MarketPlace {
         ercErcOrders[id].offeringAmount -= buyAmount;
         ercErcOrders[id].askingAmount -= saleAmount;
 
-        IERC20(order.askingToken).safeTransferFrom(
-            msg.sender,
-            order.requesterAddress,
-            saleAmount
-        );
+        IERC20(order.askingToken).safeTransferFrom(msg.sender, order.requesterAddress, saleAmount);
         IERC20(order.offeringToken).safeTransfer(msg.sender, buyAmount);
 
         emit acceptOrder(id, msg.sender, buyAmount, saleAmount);
     }
 
-    function withdrawErcErcOrder(uint id) public {
+    function withdrawErcErcOrder(uint256 id) public {
         Order memory order = ercErcOrders[id];
         require(order.requesterAddress == msg.sender);
 
-        IERC20(order.offeringToken).safeTransfer(
-            msg.sender,
-            order.offeringAmount
-        );
+        IERC20(order.offeringToken).safeTransfer(msg.sender, order.offeringAmount);
 
         delete ercErcOrders[id];
 
@@ -114,15 +89,15 @@ contract MarketPlace {
     }
 
     function getOpenOrders() external view returns (Order[] memory) {
-        uint numOpenOrders = 0;
-        for (uint i = 0; i < nextOrderId; i++) {
+        uint256 numOpenOrders = 0;
+        for (uint256 i = 0; i < nextOrderId; i++) {
             if (ercErcOrders[i].offeringAmount > 0) {
                 numOpenOrders++;
             }
         }
         Order[] memory ret = new Order[](numOpenOrders);
-        uint numPushed = 0;
-        for (uint i = 0; i < nextOrderId; i++) {
+        uint256 numPushed = 0;
+        for (uint256 i = 0; i < nextOrderId; i++) {
             if (ercErcOrders[i].offeringAmount > 0) {
                 ret[numPushed] = ercErcOrders[i];
                 numPushed++;
